snowflake_airflow/requirements.txt
apache-airflow==2.7.0
snowflake-snowpark-python==1.9.0